Create DynamicPropertyEditor.js
Literally just setScore and getScore functions but it only works for players and it's dynamic properties. Nice if you want to use scores but also store text without a database, or want to store numbers up to 8 quadrillion (with decimals). You can store integers, floats, strings, booleans, and vector3s because dynamic properties are weird.